Upload code for lesson 9
In this lesson you are shown how to download QiskitBlocks, a Minetest mod that contains puzzles in quantum circuits logic. The code I uploaded is just a different implementation of the quantum teleportation protocol that I copied from QiskitBlocks and it's in QASM. I then transfered it to Qiskit through IBM's circuit composer.